Enable the GitHub Actions Helm Release Workflow
* Enable the GitHub Actions Helm Release Workflow

Currently the releases have to be done manually and there is no Helm
Repository.

This adds automation for checking the main branch for updates in the
`Chart.yaml` file. If there is a new release the Helm Chart get built
and uploaded to the GitHub releases. Then the index file on the branch
`gh-pages` gets updates with the latest release.

There is no tagging necessary. The action will only release if there are
changes to the Helm Chart Version.
This requires the GitHub pages to be active and point to the branch
`gh-pages`.