Add ability to mount appsFromVolumes
To get an app into the Pod for metaphactory to use, we need to mount it
inside the Pod. There is currently no way to do so.

This adds an option to define `appsFromVolumes` in the values as a list
of volume definition. This could be a ConfigMap or a Secret that needs
to be deployed beforehand.

The ConfigMap needs to contain the structure of a regular metaphactory
app but since it is not possible to put `/` into a ConfigMap keys, it
needs to be escaped to `--`. The value of the key `assets--logo.svg`
will thus be mounted under `/app/${nameOfConfigMap}/assets/logo.svg`.