Add option to provide rawDatabaseConfig TTLs
Currently you can only change some of the contests of the TTL files for
default, tests, and assets repositories. This means you cannot switch to
another database type or do other modifications to the connections.

This change enables setting values that overwrite the entire TTL files
for either all or single TTL files.